第一个版本，从远端获取代理配置，并且定时进行刷新

+++ pkg/msg/msg.go
@@ -213,7 +213,7 @@ type User struct {
 
 type Frps struct {
 	Ip   string `json:"ip,omitempty"`
-	Port uint16 `json:"port,omitempty"`
+	Port int    `json:"port,omitempty"`
 }
 
 type Proxy struct {
